backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Azure SDK
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-ai-formrecognizer==3.3.2
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# OpenAI (no LangChain - we built custom agent framework)
openai>=1.10.0

# Database
pyodbc==5.0.1
redis==5.0.1

# ML and vector operations
numpy==1.26.2
tiktoken==0.5.1

# HTTP client
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
